Error handling and cleanup

+++ llama_index/vector_stores/redis.py
@@ -170,6 +170,13 @@ class RedisVectorStore(VectorStore):
         query_str = "@doc_id:{%s}" % self.tokenizer.escape(doc_id)
         # find all documents that match a doc_id
         results = self._redis_client.ft(self._index_name).search(query_str)
+        if len(results.docs) == 0:
+            # don't raise an error but warn the user that document wasn't found
+            # could be a result of eviction policy
+            _logger.warning(
+                f"Could not find document with doc_id {doc_id} in index {self._index_name}"
+            )
+            return
 
         for doc in results.docs:
             self._redis_client.delete(doc.id)
@@ -190,8 +197,14 @@ class RedisVectorStore(VectorStore):
 
         Returns:
             VectorStoreQueryResult: query result
+
+        Raises:
+            ValueError: If query.query_embedding is None.
+            redis.exceptions.RedisError: If there is an error querying the index.
+            redis.exceptions.TimeoutError: If there is a timeout querying the index.
         """
-        from redis.exceptions import ResponseError as RedisResponseError
+        from redis.exceptions import RedisError
+        from redis.exceptions import TimeoutError as RedisTimeoutError
 
         return_fields = ["id", "doc_id", "text", self._vector_key, "vector_score"]
 
@@ -212,8 +225,11 @@ class RedisVectorStore(VectorStore):
             results = self._redis_client.ft(self._index_name).search(
                 redis_query, query_params=query_params  # type: ignore
             )
-        except RedisResponseError as e:
-            _logger.error(f"Error querying index {self._index_name}: {e}")
+        except RedisTimeoutError as e:
+            _logger.error(f"Query timed out on {self._index_name}: {e}")
+            raise e
+        except RedisError as e:
+            _logger.error(f"Error querying {self._index_name}: {e}")
             raise e
 
         ids = []
@@ -239,13 +255,23 @@ class RedisVectorStore(VectorStore):
         Args:
             persist_path (str): Path to persist the vector store to. (doesn't apply)
             in_background (bool, optional): Persist in background. Defaults to True.
+
+        Raises:
+            redis.execptions.RedisError: If there is an error persisting the index to disk.
         """
-        if in_background:
-            _logger.info("Saving index to disk in background")
-            self._redis_client.bgsave()
-        else:
-            _logger.info("Saving index to disk")
-            self._redis_client.save()
+        from redis.exceptions import RedisError
+
+        try:
+            if in_background:
+                _logger.info("Saving index to disk in background")
+                self._redis_client.bgsave()
+            else:
+                _logger.info("Saving index to disk")
+                self._redis_client.save()
+
+        except RedisError as e:
+            _logger.error(f"Error saving index to disk: {e}")
+            raise e
 
     def _create_index(self) -> None:
         # should never be called outside class and hence should not raise importerror
@@ -312,34 +338,40 @@ class RedisVectorStore(VectorStore):
         returns:
             A RediSearch VectorField.
         """
+        from redis import DataError
         from redis.commands.search.field import VectorField
 
-        if algorithm.upper() == "HNSW":
-            return VectorField(
-                name,
-                "HNSW",
-                {
-                    "TYPE": datatype.upper(),
-                    "DIM": dims,
-                    "DISTANCE_METRIC": distance_metric.upper(),
-                    "INITIAL_CAP": initial_cap,
-                    "M": m,
-                    "EF_CONSTRUCTION": ef_construction,
-                    "EF_RUNTIME": ef_runtime,
-                    "EPSILON": epsilon,
-                },
-            )
-        else:
-            return VectorField(
-                name,
-                "FLAT",
-                {
-                    "TYPE": datatype.upper(),
-                    "DIM": dims,
-                    "DISTANCE_METRIC": distance_metric.upper(),
-                    "INITIAL_CAP": initial_cap,
-                    "BLOCK_SIZE": block_size,
-                },
+        try:
+            if algorithm.upper() == "HNSW":
+                return VectorField(
+                    name,
+                    "HNSW",
+                    {
+                        "TYPE": datatype.upper(),
+                        "DIM": dims,
+                        "DISTANCE_METRIC": distance_metric.upper(),
+                        "INITIAL_CAP": initial_cap,
+                        "M": m,
+                        "EF_CONSTRUCTION": ef_construction,
+                        "EF_RUNTIME": ef_runtime,
+                        "EPSILON": epsilon,
+                    },
+                )
+            else:
+                return VectorField(
+                    name,
+                    "FLAT",
+                    {
+                        "TYPE": datatype.upper(),
+                        "DIM": dims,
+                        "DISTANCE_METRIC": distance_metric.upper(),
+                        "INITIAL_CAP": initial_cap,
+                        "BLOCK_SIZE": block_size,
+                    },
+                )
+        except DataError as e:
+            raise ValueError(
+                f"Failed to create Redis index vector field with error: {e}"
             )
 
 
@@ -347,5 +379,9 @@ def cast_metadata_types(mapping: Optional[Dict[str, Any]]) -> Dict[str, str]:
     metadata = {}
     if mapping:
         for key, value in mapping.items():
-            metadata[str(key)] = str(value)
+            try:
+                metadata[str(key)] = str(value)
+            except (TypeError, ValueError) as e:
+                # warn the user and continue
+                _logger.warning("Failed to cast metadata key %s to string: %s", key, e)
     return metadata
